+++ easypost/constant.py
@@ -1,8 +1,2 @@
-import platform
-
-from .version import VERSION
-
-
 SUPPORT_EMAIL = "support@easypost.com"
-USER_AGENT = f"EasyPost/v2 PythonClient/{VERSION} Python/{platform.python_version()}"
 TIMEOUT = 60